# perf: cache prepared statements across .prepareStatement calls