… doc/pgjdbc.xml
… org/postgresql/core/v3/QueryExecutorImpl.java
… org/postgresql/core/v3/SimpleQuery.java
… org/postgresql/ds/common/BaseDataSource.java
… org/postgresql/jdbc2/AbstractJdbc2Connection.java
… org/postgresql/jdbc2/AbstractJdbc2Statement.java
… org/postgresql/jdbc2/CachedQueryCreateAction.java
… org/postgresql/jdbc2/CallableQueryKey.java
… org/postgresql/test/util/LruCacheTest.java
… org/postgresql/util/CanEstimateSize.java
… org/postgresql/util/LruCache.java
… ubenchmark/src/main/java/org/postgresql/benchmark/statement/ParseStatement.java