This allows to use server-side prepared statements when application uses the same SQL multiple times.

Major difference from existing `setPrepareThreshold` is with current fix application does not have to reference exactly the same `PreparedStatement`. When a new statement is prepared, it is checked against existing statements in the cache. If there is a match, parsing result and server-side named query (!) are reused.

Note: this fix comes on top of https://github.com/pgjdbc/pgjdbc/pull/311.

### Note: I believe the code is feature complete

Cache size is configurable. It defaults to 256 queries and 5 MiB maximum per connection.
Feel free to review, add your comments, and merge.

Some measurements:
On a simple `select ?, ?, ?...` the response time is 20-40% better (78us -> 46us for 20 binds).
Execution of cached statements allocates 50% less garbage.

### Before

```
Benchmark                                           (bindCount)  Mode  Cnt     Score      Error   Units
ParseStatement.bindExecuteFetch                               0  avgt   10    42,883 ±    0,519   us/op
ParseStatement.bindExecuteFetch:·gc.alloc.rate.norm           0  avgt   10  1160,077 ±    0,255    B/op
ParseStatement.bindExecuteFetch                               1  avgt   10    48,658 ±    0,483   us/op
ParseStatement.bindExecuteFetch:·gc.alloc.rate.norm           1  avgt   10  1600,087 ±    0,289    B/op
ParseStatement.bindExecuteFetch                              10  avgt   10    64,844 ±    0,657   us/op
ParseStatement.bindExecuteFetch:·gc.alloc.rate.norm          10  avgt   10  3896,114 ±    0,379    B/op
ParseStatement.bindExecuteFetch                              20  avgt   10    78,440 ±    0,772   us/op
ParseStatement.bindExecuteFetch:·gc.alloc.rate.norm          20  avgt   10  6576,138 ±    0,458    B/op
```

### After

```
Benchmark                                           (bindCount)  Mode  Cnt     Score     Error   Units
ParseStatement.bindExecuteFetch                               0  avgt   10    32,450 ±   0,195   us/op
ParseStatement.bindExecuteFetch:·gc.alloc.rate.norm           0  avgt   10   608,057 ±   0,190    B/op
ParseStatement.bindExecuteFetch                               1  avgt   10    34,650 ±   0,907   us/op
ParseStatement.bindExecuteFetch:·gc.alloc.rate.norm           1  avgt   10   792,061 ±   0,201    B/op
ParseStatement.bindExecuteFetch                              10  avgt   10    41,341 ±   0,315   us/op
ParseStatement.bindExecuteFetch:·gc.alloc.rate.norm          10  avgt   10  1464,073 ±   0,241    B/op
ParseStatement.bindExecuteFetch                              20  avgt   10    46,994 ±   0,352   us/op
ParseStatement.bindExecuteFetch:·gc.alloc.rate.norm          20  avgt   10  2224,083 ±   0,275    B/op
```

Pathological case when `prepareStatement` is called with random sql texts.
1) Test passes, no out of memory happens
2) Response time is intact
3) Memory allocation is a bit higher than before 4496 -> 4560

### Before

```
Benchmark                                                         (bindCount)  (unique)  Mode  Cnt     Score      Error   Units
ParseStatement.bindExecuteFetch                                            10      true  avgt   10    64,710 ±    0,495   us/op
ParseStatement.bindExecuteFetch:·gc.alloc.rate.norm                        10      true  avgt   10  4496,114 ±    0,378    B/op
```

### After

```
Benchmark                                                         (bindCount)  (unique)  Mode  Cnt     Score      Error   Units
ParseStatement.bindExecuteFetch                                            10      true  avgt   10    65,289 ±    0,568   us/op
ParseStatement.bindExecuteFetch:·gc.alloc.rate.norm                        10      true  avgt   10  4560,115 ±    0,384    B/op
```