@sehrope 2015-06-28
Haven't had a chance to go through this yet (skimming through it now) but overall I like the idea.

I wonder if it'd be worth it to have this type of caching done at the Driver level so that it can be shared across multiple connections. That way a pool of connections would share a single cache, even as connections are added/removed from the pool (ex: pool expiration). For that to happen it'd at least have to be thread safe. I don't think there's a direct concurrent replacement for LinkedHashMap so that'd need to be sorted out. Also, we may need separate caches per connection URL as the parsing of a particular query is dependent on connection level settings.

It's a trade off of code complexity vs a wider gain. For a pool of size N using K memory, it'd be (N-1) x K savings so I guess it'd be worth it for a large pool or diverse workload. If it's too complicated (either to implement or we just decide it's too risky) then I wouldn't want to hold off the version of the patch that is only done within a connection. Just something to consider...