@sehrope 2015-06-28
One more thought ... how bout an option to not cache queries that are larger than a given size? In most real-world applications the majority of queries are relatively small though every now and then there will be some dynamic SQL that is huge. Usually they're large dynamically generated reporting queries. Although they'd eventually get washed out of the cache, it's pointless to cache them as they're one offs.

At the very least we shouldn't bother trying to cache things that are greater than the max size of the cache. Otherwise one gigantic query would evict all the prior smaller queries (and itself!).