@sehrope 2015-06-28
> Not sure how you do this as server side statements are per connection ??

I was only thinking about caching the driver side parsing of the sql String. The idea being that there is no need to repeatedly parse the same sql as it'll always have the same bind var positions and native representation (i.e. `SELECT ?` always goes to `SELECT $1`). That would improve perf (less parsing) and GC (less parsing means less garbage).

I haven't really gone through the patch and didn't realize that it caches the server side prepared statements and is intended to reuse those as well. Thinking that through a bit, I wonder if the two things can be split out. There's no reason you have to use server side prepared statements to benefit from driver side parse caching right?