@davecramer 2015-06-29
On 29 June 2015 at 05:25, Vladimir Sitnikov <notifications@github.com>
wrote:

> I don't think there's a direct concurrent replacement for LinkedHashMap so
> that'd need to be sorted out
>
> It is not in JDK, but it is does support concurrency and expiration.
> Here it is: https://github.com/ben-manes/concurrentlinkedhashmap
> Not sure if we need that, but your point is valid.
>
> At the very least we shouldn't bother trying to cache things that are
> greater than the max size of the cache. Otherwise one gigantic query would
> evict all the prior smaller queries (and itself!).
>
> I've filed #344 <https://github.com/pgjdbc/pgjdbc/issues/344> to cover
> that.
>
> Thinking that through a bit, I wonder if the two things can be split out.
> There's no reason you have to use server side prepared statements to
> benefit from driver side parse caching right?
>
> I did intend to split those two items, however I prefer filing another PR
> to cover "inter-connection" caching of query texts.
>
> Here is the issue: #345 <https://github.com/pgjdbc/pgjdbc/issues/345>
>
> @davecramer <https://github.com/davecramer>, technically I did not
> expected to get PR merged in while still having non-resolved comments. I'll
> file another "cleanup" PR.
>

I'd like to release snapshots. I can wait to release a version until it is
cleaned up. Note that my merging it evoked constructive criticism.