@vlsi 2015-06-29
>I can wait to release a version until it is cleaned up.

Here's the cleanup PR: https://github.com/pgjdbc/pgjdbc/pull/346