@davecramer 2015-06-29
Merged, thanks!

Dave Cramer

On 29 June 2015 at 08:20, Vladimir Sitnikov <notifications@github.com>
wrote:

> I can wait to release a version until it is cleaned up.
>
> Here's the cleanup PR: #346 <https://github.com/pgjdbc/pgjdbc/pull/346>
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/pgjdbc/pgjdbc/pull/319#issuecomment-116633436>.
>